set correct SE Linux context on created directories and files
Without this change, resolvconf is incompatible with SE Linux:

  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=740685

I've written SE Linux policy to fix the above bug, but we also need 2 minor
changes to resolvconf.

A file named /usr/lib/tmpfiles.d/resolvconf.conf with appopriate contents
causes systemd to create the temporary directories and files with the
correct SE Linux context.  It also removes the need for making a
directory in the ExecStartPre section of /lib/systemd/system/resolvconf.service.

This change makes it work correctly on SysVInit.  On systems that don't run
SE Linux it has no effect.

Closes: #850783